first big refactoring and folder re-structure
