+++ app/config.json
@@ -1,0 +1,11 @@
+{
+    "data_directory" : "./app/_data",
+    "vectordb_directory": "{data_directory}/mychroma_db",
+    "summaries_directory": "{data_directory}/summaries",
+    "files_directory" : "{data_directory}/files",
+    "chat_history_file" : "{data_directory}/chat_history.json",
+    "qa_tracker_file" : "{data_directory}/qa_tracker.json",
+    "embedding_model": "text-embedding-3-small",
+    "llm": "gpt-3.5-turbo",
+    "n_sources_to_display": 3
+}
