+++ app/models.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class AskQuestionRequest(BaseModel):
+    question: str
+    owner: str
