+++ pyproject.toml
@@ -13,10 +13,10 @@ langchain = "^0.1.14"
 langchain-openai = "^0.1.1"
 docx2txt = "^0.8"
 pypdf = "^4.1.0"
-streamlit = "^1.33.0"
 tiktoken = "^0.6.0"
 python-multipart = "^0.0.9"
 chromadb = "0.4.14"
+pydantic-settings = "^2.3.2"
 
 
 [build-system]
